+++ src/Common/Resource.fs
@@ -3,10 +3,15 @@ namespace KafkaApplication
 module ResourceChecker =
     open Kafka
     open Metrics
+    open ServiceIdentification
 
     let updateResourceStatusOnCheck instance (BrokerList brokerList) kafkaChecker: Kafka.Checker =
         let kafkaClusterResource = ResourceAvailability.createFromStrings "kafka_cluster" brokerList brokerList Audience.Sys
-        let kafkaTopicResource (StreamName topic) = ResourceAvailability.createFromStrings "kafka_topic" topic brokerList Audience.Sys
+        let kafkaTopicResource = function
+            | (StreamName topic) ->
+                ResourceAvailability.createFromStrings "kafka_topic" topic brokerList Audience.Sys
+            | Instance topic ->
+                ResourceAvailability.createForServiceFromStrings "kafka_topic" (topic |> Instance.concat "-") brokerList topic Audience.Sys
 
         let updateResourceStatus (resource: ResourceAvailability) = function
             | true -> ResourceAvailability.enable instance resource |> ignore